+++ jobbergate-api/tests/apps/clusters/test_routers.py
@@ -125,3 +125,39 @@ class TestListClusterStatus:
         response = await client.get("/jobbergate/clusters/status")
 
         assert response.status_code == status.HTTP_403_FORBIDDEN
+
+
+class TestGetClusterStatus:
+    @pytest.mark.parametrize("permission", (Permissions.ADMIN, Permissions.CLUSTERS_READ))
+    @pytest.mark.parametrize("client_id", ("client-1", "client-2", "client-3"))
+    async def test_get_cluster_status_by_client_id__not_found(
+        self, permission, client_id, client, inject_security_header, synth_session
+    ):
+        inject_security_header("who@cares.com", permission)
+
+        response = await client.get(f"/jobbergate/clusters/status/{client_id}")
+        assert response.status_code == status.HTTP_404_NOT_FOUND
+
+    @pytest.mark.parametrize("permission", (Permissions.ADMIN, Permissions.CLUSTERS_READ))
+    @pytest.mark.parametrize("client_id", ("client-1", "client-2", "client-3"))
+    async def test_get_cluster_status_by_client_id__found(
+        self, permission, client_id, client, inject_security_header, synth_session
+    ):
+        inject_security_header("who@cares.com", permission)
+
+        cluster_status = ClusterStatus(
+            client_id=client_id, interval=10, last_reported=pendulum.datetime(2023, 1, 1)
+        )
+        with pendulum.travel_to(pendulum.datetime(2023, 1, 1), freeze=True):
+            synth_session.add(cluster_status)
+            response = await client.get(f"/jobbergate/clusters/status/{client_id}")
+
+        assert response.status_code == status.HTTP_200_OK
+        assert response.json() == {
+            "client_id": client_id,
+            "interval": 10,
+            "last_reported": "2023-01-01T00:00:00Z",
+            "created_at": "2023-01-01T00:00:00Z",
+            "updated_at": "2023-01-01T00:00:00Z",
+            "is_healthy": True,
+        }